refactor(domain): restructure product, category, and review entities and repos
- Moved `Product`, `Category`, and `Review` entities into domain-based subpackages for better modularity
- Organized repository interfaces (`ProductRepository`, `CategoryRepository`, `ReviewRepository`) under corresponding `repo` packages
