+++ src/main/java/com/eric/ecommerce_product_service/category/repo/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.eric.ecommerce_product_service.repositories;
+package com.eric.ecommerce_product_service.category.repo;
 
-import com.eric.ecommerce_product_service.entities.Category;
+import com.eric.ecommerce_product_service.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
